extract service.ResolveURLQueryPathInt64SliceFieldWithLimit() from items.idsFromRequest(), rename/split files of the service package
